requirements/test.txt
# Test dependencies go here.
-r base.txt
coverage==3.7.1
django-extratest>=2.0
